refactor: rewrite storages
