+++ task-1/src/services/index.ts
@@ -1,1 +1,1 @@
-export * from './storage';
+export * from './storages';
