+++ task-1/src/services/storages/index.ts
@@ -1,0 +1,1 @@
+export * from './terms-storage';
