+++ task-1/src/services/storages/storage/index.ts
@@ -1,0 +1,1 @@
+export * from './storage';
