+++ task-1/src/services/storages/terms-storage.ts
@@ -1,0 +1,10 @@
+import { Storage } from './storage';
+
+type TermsStorageKeys = 'searchTerms';
+
+const TermsStorage = new Storage<TermsStorageKeys>(
+  localStorage
+);
+Object.freeze(TermsStorage);
+
+export { TermsStorage };
